Combined all event manager code into one, and commented out a couple of lines of code from some sort of scene transition
