+++ Assets/Building-MG.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
+guid: 4d7ad6a58308efc48a92d83aaa847f3e
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
