+++ Assets/Building-MG/Assets.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: e0d9be6f568183d42af89beab6592967
+guid: 8b701bd000c54a647a72801db0ec942f
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
