--- Assets/Building-MG/Assets/EventManagerBuild.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4cd6a97043b415246a8bd0998744847a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
