--- Assets/Car-MG/EventManager2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85aa2d5bb0f4a944fac9f08500c642c0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
