+++ Assets/Car-MG/WriteWin.cs
@@ -18,11 +18,11 @@ public class WriteWin : MonoBehaviour
     }
     void OnEnable()
     {
-        EventManager2.WinEvent += WriteWinText; //adds player crash to crashEvent
+        EventManagerMain.WinEvent += WriteWinText; //adds player crash to crashEvent
     }
 
     void OnDisable()
     {
-        EventManager2.WinEvent -= WriteWinText;//removes player crash to crashEvent
+        EventManagerMain.WinEvent -= WriteWinText;//removes player crash to crashEvent
     }
 }
