--- Assets/Dressing-MG/Scripts/EventManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f51dd430b0b0c16479fe31bee6f10035
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
