+++ Assets/Dressing-MG/Scripts/Shirt/ShirtEvent.cs
@@ -11,11 +11,11 @@ public class ShirtEvent : MonoBehaviour
 
     private void OnEnable()
     {
-        EventManager.ShirtOnTouched += Shirt;
+        EventManagerMain.ShirtOnTouched += Shirt;
     }
 
     private void OnDisable()
     {
-        EventManager.ShirtOnTouched -= Shirt;
+        EventManagerMain.ShirtOnTouched -= Shirt;
     }
 }
