--- Assets/Murder-MG/Scripts/MurderEventManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: eecab13e79117934eb08ca2edf54a813
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
