+++ Assets/Scenes/Building.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e0d9be6f568183d42af89beab6592967
+guid: 81dca8ce95434ad4aa9d9309f18b1d04
 DefaultImporter:
   externalObjects: {}
   userData:
